+++ Source/DSP/SoftClipper.h
@@ -12,7 +12,7 @@ class SoftClipper
 public:
     SoftClipper();
     void update(float gain);
-    void process(juce::AudioBuffer<float>& buffer, int numSamples);
+    void process(juce::AudioBuffer<float>& buffer, size_t numSamples);
 private:
     float _applyGain(float x) const;
     float _applyCubicNonlinearity(float x);
